finish unit tests

+++ src/services/DeleteDeliveryService.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from 'tsyringe';
 
 import IDeliveryRepository from '../repositories/models/IDeliveryRepository';
+import AppError from '../helpers/AppError';
 
 interface IRequest {
   id: number;
@@ -14,11 +15,16 @@ class DeleteDeliveryService {
   ) {}
 
   public async execute(input: IRequest): Promise<void> {
-    const deliveries = await this.deliveryRepository.delete(input.id);
-    console.log(
-      'DeleteDeliveryService :: deliveries :: ',
-      JSON.stringify(deliveries),
+    const deliveryAlreadyExists = await this.deliveryRepository.findById(
+      input.id,
     );
+
+    if (!deliveryAlreadyExists) {
+      throw new AppError('The Delivery id provided does not exist');
+    }
+
+    await this.deliveryRepository.delete(input.id);
+    console.log('DeleteDeliveryService :: delivery deleted with success');
   }
 }
